Implemented removing folder.

+++ documents-keeper-ui/src/app/documents-keeper/folder-configuration/field-list/field-list.component.ts
@@ -1,5 +1,5 @@
 import { Component, OnInit } from '@angular/core';
-import { ActivatedRoute } from '@angular/router';
+import { ActivatedRoute, Router } from '@angular/router';
 import { Observable } from 'rxjs/Observable';
 import { Field } from 'app/model/field.model';
 import { FieldRepository } from 'app/model/repositories/field.repository';
@@ -8,7 +8,8 @@ import { Folder } from 'app/model/folder.model';
 
 @Component({
     moduleId: module.id,
-    templateUrl: 'field-list.component.html'
+    templateUrl: 'field-list.component.html',
+    styleUrls: ['field-list.component.scss']
 })
 export class FieldListComponent implements OnInit {
     public folderId: string;
@@ -23,15 +24,19 @@ export class FieldListComponent implements OnInit {
     constructor(
         private readonly fieldRepository: FieldRepository,
         private readonly folderRepository: FolderRepository,
-        private readonly activeRoute: ActivatedRoute) {
+        private readonly activeRoute: ActivatedRoute,
+        private readonly router: Router) {
             this.folderId = activeRoute.snapshot.params['folderId'];
     }
 
     public removeField(fieldId: number) {
-        this.fieldRepository.removeField(fieldId)
-            .subscribe(data => {
-                // this.fields.removeFromArray
-                console.log('Field was successfully removed.');
+        if(confirm('Do you really want to delete field? You will lost all of your field values.'))
+        {
+            this.fieldRepository.removeField(fieldId)
+                .subscribe(data => {
+                    this.router
+                        .navigateByUrl(`/folders/${this.folderId}/configuration/fields`);
             });
+        }
     }
 }
